requirements.txt
asgiref==3.9.1
Django==4.2.23
mysqlclient==2.2.7
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow>=10.3.0
torch
torchvision
djangorestframework-simplejwt
python-decouple
requests
